Fix data type for variables in B. Memo and Momo problem; add C. Next Alphabet implementation

+++ B_Memo_and_Momo.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main() {
 
-    int a ,b,k;
+    long long a ,b,k;
     cin >> a>>b>>k;
 
     if ( a%k == 0 && b%k == 0)
